+++ src/Exceptions/InvalidCallOrderException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace AliYavari\IranPayment\Exceptions;
 
-use Illuminate\Support\Arr;
+use AliYavari\IranPayment\Enums\ApiMethod;
 use LogicException;
 
 /**
@@ -13,12 +13,30 @@ use LogicException;
 final class InvalidCallOrderException extends LogicException
 {
     /**
-     * @param  array<string>  $requiredMethods
+     * @param  array<ApiMethod>  $requiredMethods
      */
-    public static function make(string $attemptedMethod, array $requiredMethods): self
+    public static function mustBeCalledAfter(string $attemptedMethod, array $requiredMethods): self
     {
         return new self(
-            sprintf('Cannot call "%s()" before calling one of the following methods: "%s".', $attemptedMethod, Arr::join($requiredMethods, ', '))
+            sprintf('Cannot call "%s()" before calling one of the following methods: "%s".', $attemptedMethod, self::joinMethods($requiredMethods))
         );
+    }
+
+    /**
+     * @param  array<ApiMethod>  $blockingMethods
+     */
+    public static function mustBeCalledBefore(string $attemptedMethod, array $blockingMethods): self
+    {
+        return new self(
+            sprintf('Cannot call "%s()" after calling one of the following methods: "%s".', $attemptedMethod, self::joinMethods($blockingMethods))
+        );
+    }
+
+    /**
+     * @param  array<ApiMethod>  $methods
+     */
+    private static function joinMethods(array $methods): string
+    {
+        return collect($methods)->pluck('value')->join(', ');
     }
 }