+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Fixed
+
+- Reading `getRefNumber()` and `getCardNumber()` after a manual `reverse()` call throws an `InvalidCallOrderException`.
+
 ## [2.0.1] - 2026-08-22
 
 ### Fixed
@@ -51,6 +57,7 @@ Supported Gateways:
 - Zarinpal
 - IDPay
 
+[unreleased]: https://github.com/amyavari/iran-payment-laravel/compare/v2.0.1...HEAD
 [2.0.1]: https://github.com/amyavari/iran-payment-laravel/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/amyavari/iran-payment-laravel/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/amyavari/iran-payment-laravel/compare/v1.0.0...v1.1.0